% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hooks.R
\name{chunk_head}
\alias{chunk_head}
\title{Print code chunk headers}
\usage{
chunk_head()
}
\description{
View the code chunk "asis" along with the chunk output.
}
\details{
To run this hook, call the \code{chunk_head()} function and then you can
either print all source code as verbatim chunks. This function performs a similar
function to \code{source_verbatim()}, but this version will keep the code within
the shaded code environment.

\preformatted{
```{r, chunk_head = TRUE}
print(mtcars)
```
}
}
\examples{
chunk_head()

}
