% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hooks-package.R
\docType{package}
\name{hooks-package}
\alias{hooks}
\alias{hooks-package}
\title{Provide useful knitr hooks to extend the functionality of knitr, Rmarkdown
and bookdown.}
\description{
Provide useful knitr hooks to extend the functionality of knitr,
    Rmarkdown and bookdown.
}
\author{
\strong{Maintainer}: Nathan Eastwood \email{nathan.eastwood@icloud.com}

Other contributors:
\itemize{
  \item Michael Harper [contributor]
}

}
