% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hooks.R
\name{output_lines}
\alias{output_lines}
\title{Print certain output lines}
\usage{
output_lines()
}
\description{
Print only lines of the R output that the user specifies. This works
similarly to how \code{echo} works for printing R code.
}
\details{
To run this hook, call the \code{output_lines()} function and then you can
either set a global maximum height

\code{
knitr::opts_chunk$set(output_lines = 1:10)
}

or you can specify it on a chunk by chunk basis

\preformatted{
```{r output_lines = 3:15}
print(mtcars)
```
}

Several options are available
\describe{
  \item{output_lines = n}{prints lines 1:n ...}
  \item{output_lines = 1:n}{prints lines 1:n ...}
  \item{output_lines = 3:15}{prints lines ... 3:15 ...}
  \item{output_lines = -(1:8)}{removes lines 1:8 and prints ... 9:n ...}
}
Note, there is no allowance for anything but a consecutive range of lines
}
\examples{
output_lines()

}
