% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hooks.R
\name{output_max_height}
\alias{output_max_height}
\title{Add a scrollbar to R output}
\usage{
output_max_height()
}
\description{
Determine a maximum height that R console output should be and if output is
larger than that value, make it scrollable.
}
\details{
To run this hook, call the \code{output_max_height()} function and then you
can either set a global maximum height

\code{
knitr::opts_chunk$set(output_max_height = "300px")
}

or you can specify it on a chunk by chunk basis

\preformatted{
```{r output_max_height = "300px"}
print(mtcars)
```
}
}
\examples{
output_max_height()

}
