% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hooks.R
\name{source_verbatim}
\alias{source_verbatim}
\title{Print verbatim code chunks}
\usage{
source_verbatim()
}
\description{
View the code chunk "asis" along with the chunk output.
}
\details{
To run this hook, call the \code{source_verbatim()} function and then you can
either print all source code as verbatim chunks

\code{
knitr::opts_chunk$set(source_verbatim = TRUE)
}

or you can print it on a chunk by chunk basis

\preformatted{
```{r source_verbatim = TRUE}
print(mtcars)
```
}

If you would like to only print the source code as verbatim and not have the
code evaluated, use \code{eval = FALSE}, i.e.

\preformatted{
```{r, eval = FALSE, source_verbatim = TRUE}
print(mtcars)
```
}

The resulting output will look like this.

\preformatted{
```{r}
print(mtcars)
```
}
}
\examples{
source_verbatim()

}
